Подготовил ручку и скелет метода, выделил шаблон результата валидации
